Moved upgrade logic to a separate class.
