+++ VoiceAttackResponder/VoiceAttackPlugin.cs
@@ -151,23 +151,23 @@ namespace EddiVoiceAttackResponder
                 }
 
                 // Display instance information if available
-                if (App.UpgradeRequired)
+                if (EddiUpgrader.UpgradeRequired)
                 {
                     vaProxy.WriteToLog("Please shut down VoiceAttack and run EDDI standalone to upgrade", "red");
                     string msg = Properties.VoiceAttack.run_eddi_standalone;
                     SpeechService.Instance.Say(((ShipMonitor)EDDI.Instance.ObtainMonitor("Ship monitor")).GetCurrentShip(), msg, 0);
                 }
-                else if (App.UpgradeAvailable)
+                else if (EddiUpgrader.UpgradeAvailable)
                 {
                     vaProxy.WriteToLog("Please shut down VoiceAttack and run EDDI standalone to upgrade", "orange");
                     string msg = Properties.VoiceAttack.run_eddi_standalone;
                     SpeechService.Instance.Say(((ShipMonitor)EDDI.Instance.ObtainMonitor("Ship monitor")).GetCurrentShip(), msg, 0);
                 }
 
-                if (App.Motd != null)
+                if (EddiUpgrader.Motd != null)
                 {
-                    vaProxy.WriteToLog("Message from EDDI: " + App.Motd, "black");
-                    string msg = String.Format(Eddi.Properties.EddiResources.msg_from_eddi, App.Motd);
+                    vaProxy.WriteToLog("Message from EDDI: " + EddiUpgrader.Motd, "black");
+                    string msg = String.Format(Eddi.Properties.EddiResources.msg_from_eddi, EddiUpgrader.Motd);
                     SpeechService.Instance.Say(((ShipMonitor)EDDI.Instance.ObtainMonitor("Ship monitor")).GetCurrentShip(), msg, 0);
                 }
